requirements.txt
autogenstudio==0.0.32a0
boto3==1.34.28
beautifulsoup4==4.12.3
bleach[css]==6.1.0 
celery==5.3.6
channels==4.0.0
channels-redis==4.2.0
daphne==4.0.0
Django==5.0.1
django-cors-headers==4.3.1
django-countries==7.5.1
django-extensions==3.2.3
django-filter==23.5
django-on-heroku==1.1.2
django-rest-passwordreset==1.3.0
django-rest-swagger==2.2.0
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dj-database-url==2.1.0
# docx2pdf==0.1.8  # only used in a window/Mac environment
drf-spectacular==0.27.1
drf-yasg==1.21.7
factory_boy==3.3.0
google-apps-meet==0.1.1
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-cloud-pubsub==2.19.0
google-cloud-storage==2.14.0
gunicorn==21.2.0
langchain==0.1.4
llama-index==0.9.38
markdown==3.5.2
more-itertools==10.2.0
numpy==1.26.3
openai==1.10.0
open-interpreter==0.2.0
pandas==2.2.0
Pillow==10.2.0
pinecone-client==3.0.2
playwright==1.41.1
pyautogen==0.2.8
pymongo==4.6.1
pyspellchecker==0.8.1
python-decouple==3.8
python-dotenv==1.0.1
python-Levenshtein==0.23.0
scikit-learn==1.4.0
tabula-py==2.9.0
textblob==0.17.1
textstat==0.7.3
tiktoken<0.5.0  # specifically required for open-interpreter at the moment
unstructured[pdf]
